Interval notes in the scene play the sound
Using an instance of the Note class as the scene elements of the generated interval with the correct positioning. User can hear the sound of the notes in the interval and see the playing note with a different color in the middle of the musical stave.
Mouse is the raycaster origin